requirements.txt
PyQt5
Pillow
keyboard